fix spelling error
