+++ lib/ResolverFactory.js
@@ -23,7 +23,7 @@ const ImportsFieldPlugin = require("./ImportsFieldPlugin");
 const JoinRequestPartPlugin = require("./JoinRequestPartPlugin");
 const JoinRequestPlugin = require("./JoinRequestPlugin");
 const MainFieldPlugin = require("./MainFieldPlugin");
-const ModulesInHierachicDirectoriesPlugin = require("./ModulesInHierachicDirectoriesPlugin");
+const ModulesInHierarchicalDirectoriesPlugin = require("./ModulesInHierarchicalDirectoriesPlugin");
 const ModulesInRootPlugin = require("./ModulesInRootPlugin");
 const NextPlugin = require("./NextPlugin");
 const ParsePlugin = require("./ParsePlugin");
@@ -406,7 +406,7 @@ exports.createResolver = function (options) {
 		if (Array.isArray(item)) {
 			if (item.includes("node_modules") && pnpApi) {
 				plugins.push(
-					new ModulesInHierachicDirectoriesPlugin(
+					new ModulesInHierarchicalDirectoriesPlugin(
 						"raw-module",
 						item.filter(i => i !== "node_modules"),
 						"module"
@@ -417,7 +417,11 @@ exports.createResolver = function (options) {
 				);
 			} else {
 				plugins.push(
-					new ModulesInHierachicDirectoriesPlugin("raw-module", item, "module")
+					new ModulesInHierarchicalDirectoriesPlugin(
+						"raw-module",
+						item,
+						"module"
+					)
 				);
 			}
 		} else {
